refactor: merge type files
